+++ java/src/main/java/com/techelevator/application/Drink.java
@@ -1,13 +1,15 @@
 package com.techelevator.application;
 
+import java.math.BigDecimal;
+
 public class Drink extends VendingItems {
 
-    public Drink(String slotNo, String nameOfProduct, int price) {
+    public Drink(String slotNo, String nameOfProduct, BigDecimal price) {
         super(slotNo, nameOfProduct, price);
     }
 
     @Override
-    public String dispenseMessage(boolean isDiscount, int remainingBalance) {
+    public String dispenseMessage(boolean isDiscount, BigDecimal remainingBalance) {
         System.out.println(getNameOfProduct() + " " + getPrice(isDiscount) + " " + remainingBalance);
         System.out.println("Drinky, Drinky, Slurp Slurp!");
         return "Drinky, Drinky, Slurp Slurp!";
